Splash screen support
Modified ipu_common.c since we have changed the base frequency of the
PFD0, therefore we need to make the ipu driver aware of it to get
correct calculations for the lvds parameters

The frequency of PFD0 was changed to 280MHz to obtain 40MHz with an
integer divisor.

Beware the removal of ipu2 from the device tree since the mxcfb_probe()
function changes the configuration of the IPU depending of the number
of calls ( via the global variable g_dp_in_use). Therefore the first
call to mxcfb_probe() from the console init code setup the IPU1 using
DIO0, the second call from the splash_display() function assume that
the last ipu defined int the DT is the one the display is connected to
and reconfigure the IPU using CSI as display port.

So we disable the IPU2 in the device tree ( since we're not using it
anyway) to prevend the second call to mxcfb_probe()